+++ src/domain/review/reviews.module.ts
@@ -1,13 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ShortReviewController } from './controller/short-review.controller';
 import { ShortReviewService } from './service/short-review.service';
-import {
-  LongReviewRepository,
-  ShortReviewRepository,
-} from './reviews.repository';
 import { LongReviewService } from './service/long-review.service';
 import { PrismaModule } from '../../global/config/prisma/prisma.module';
 import { LongReviewController } from './controller/long-review.controller';
+import { ShortReviewRepository } from './repository/short-review.repository';
+import { LongReviewRepository } from './repository/long-review.repository';
 
 @Module({
   imports: [PrismaModule],
@@ -18,6 +16,6 @@ import { LongReviewController } from './controller/long-review.controller';
     ShortReviewService,
     LongReviewService,
   ],
-  exports: [],
+  exports: [ShortReviewService, LongReviewService],
 })
 export class ReviewsModule {}
